requirements.txt
spotipy
spotipy-anon
plexapi
thefuzz
python-dotenv
rapidfuzz
spotdl
ytmusicapi
musicbrainzngs
mutagen
fastapi
uvicorn
